requirements.txt
django-crispy-forms
crispy-bootstrap4
python-decouple
Pillow
